# backend/src/main/java/com/cputhome/admin/AccreditationRequest.java
package com.cputhome.admin;

/* Explicit accreditation toggle; verification remains an independent decision. */
public record AccreditationRequest(@jakarta.validation.constraints.NotNull Boolean accredited) {}