# docs/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cput home api",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/listings": {
      "get": {
        "tags": [
          "listing-controller"
        ],
        "operationId": "browse",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "campus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "availableBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "emergency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "onCampus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "nsfas",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "gender",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "amenity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "priority"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseListingCardDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "listing-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListingCardDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{id}/enable": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "enable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{id}/disable": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "disable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/listings/{id}/unpublish": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "unpublish",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/listings/{id}/reject": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "reject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectListingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListingCardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/listings/{id}/approve": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "approve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListingCardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/{id}": {
      "get": {
        "tags": [
          "listing-controller"
        ],
        "operationId": "detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListingCardDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "listing-controller"
        ],
        "operationId": "deactivate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "listing-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ListingCardDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/providers/{id}/verification": {
      "patch": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "verify",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/providers/{id}/accreditation": {
      "patch": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "accredit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccreditationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "users",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseUserSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/providers": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "providers",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/listings": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "listings",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingCardDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "delete": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "removeUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateListingRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "campus": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "address": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          },
          "onCampus": {
            "type": "boolean"
          },
          "nsfas": {
            "type": "boolean"
          },
          "gender": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 60,
              "minLength": 0
            }
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 500,
              "minLength": 0
            }
          },
          "houseRules": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "shuttle": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "utilities": {
            "type": "number",
            "minimum": 0
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomSpec"
            },
            "maxItems": 2147483647,
            "minItems": 1
          }
        },
        "required": [
          "campus",
          "description",
          "location",
          "rooms",
          "title"
        ]
      },
      "RoomSpec": {
        "type": "object",
        "properties": {
          "roomType": {
            "type": "string",
            "minLength": 1
          },
          "monthlyRent": {
            "type": "number",
            "minimum": 0.01
          },
          "deposit": {
            "type": "number",
            "minimum": 0
          },
          "beds": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "available": {
            "type": "boolean"
          },
          "availableDate": {
            "type": "string",
            "format": "date"
          },
          "emergency": {
            "type": "boolean"
          }
        },
        "required": [
          "monthlyRent",
          "roomType"
        ]
      },
      "ListingCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "owner": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "rent": {
            "type": "number"
          },
          "location": {
            "type": "string"
          },
          "campus": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          },
          "availableDate": {
            "type": "string",
            "format": "date"
          },
          "emergency": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "onCampus": {
            "type": "boolean"
          },
          "nsfas": {
            "type": "boolean"
          },
          "gender": {
            "type": "string"
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "beds": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string"
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deposit": {
            "type": "number"
          },
          "utilities": {
            "type": "number"
          },
          "houseRules": {
            "type": "string"
          },
          "shuttle": {
            "type": "string"
          },
          "desc": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "sample": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "active": {
            "type": "boolean"
          },
          "published": {
            "type": "boolean"
          },
          "rejectionReason": {
            "type": "string"
          }
        }
      },
...
      "UserSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "RejectListingRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        }
      },
      "UpdateListingRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 150,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "campus": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "address": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          },
          "onCampus": {
            "type": "boolean"
          },
          "nsfas": {
            "type": "boolean"
          },
          "gender": {
            "type": "string",
            "maxLength": 20,
            "minLength": 0
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 60,
              "minLength": 0
            }
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 500,
              "minLength": 0
            }
          },
          "houseRules": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "shuttle": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          },
          "utilities": {
            "type": "number",
            "minimum": 0
          },
          "active": {
            "type": "boolean"
          },
          "available": {
            "type": "boolean"
          },
          "price": {
            "type": "number",
            "minimum": 0.01
          },
          "availableDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "VerifyProviderRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "VERIFIED|REJECTED"
          }
        },
        "required": [
          "status"
        ]
      },
      "ProviderResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "verificationStatus": {
            "type": "string"
          },
          "accreditation": {
            "type": "boolean"
          },
          "registrationNumber": {
            "type": "string"
          }
        }
      },
      "AccreditationRequest": {
        "type": "object",
        "properties": {
          "accredited": {
            "type": "boolean"
          }
        },
        "required": [
          "accredited"
        ]
      },
      "PageResponseListingCardDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListingCardDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
      "PageResponseUserSummary": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}